requirements.txt
Pillow>=9.5.0
tensorflow>=2.13.0
numpy>=1.24.3
ImageHash>=4.3.1
scikit-image>=0.21.0